+++ src/Line.h
@@ -10,9 +10,9 @@ using namespace std;
 class Line
 {
 public:
-	int A;
-	int B;
-	int C;//直线一般式方程Ax+By+C=0;
+	double A;
+	double B;
+	double C;//直线一般式方程Ax+By+C=0;
 	Line(int x0, int y0, int x1, int y1);
 	bool isParallel(Line line);//判断是否平行
 	pair<double, double> findIntersection(Line line) ;
